fix(toast): improve Toast structure after review
