--- src/components/providers/ToastContext.tsx
@@ -1,8 +1,0 @@
-'use client';
-
-import { createContext } from 'react';
-import { ToastContextProps } from '@/src/@types/toast.types';
-
-export const ToastContext = createContext<ToastContextProps | undefined>(
-    undefined
-);
