--- src/composables/useToast.ts
@@ -1,12 +1,0 @@
-'use client';
-
-import { ToastContext } from './../components/providers/ToastContext';
-import { useContext } from 'react';
-
-export const useToast = () => {
-    const context = useContext(ToastContext);
-    if (!context) {
-        throw new Error('useToast must be used within a ToastProvider');
-    }
-    return context;
-};
