+++ src/utils/cn.ts
@@ -2,5 +2,5 @@ import { ClassValue, clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
 
 export function cn(...inputs: ClassValue[]) {
-    return twMerge(clsx(...inputs));
+  return twMerge(clsx(...inputs));
 }
